refactor: extract shared test fixtures for hook tests (#90)
- Create tests/hooks/_fixtures.ts with shared helpers:
  userLine, interruptLine, assistantLine, setupEnv, cleanupEnv,
  writeTranscript, makeTmpDir
- Migrate pre-compact.test.ts to use shared fixtures
- Migrate session-end-transcript.test.ts to use shared fixtures
- Fix configPath missing random suffix (parallel test collision risk)
